NoSQL/RabbitMQ/requirements.txt
pika~=1.3.2
mongoengine~=0.29.1
python-dotenv~=1.0.1
Faker~=33.1.0